Branch for KSyntaxHighlighting::Theme integration
The idea of this branch is to provide a config widget for
KSyntaxHighlighting::Themes that matches the one that currently
exists to modify KateSchemas: Editor colors, Default text styles,
and Highlighting text styles.

The approach taken is to just pass a QJsonObject that represents
a full .theme file to the models. This way, the TextStyleModel etc
act as model for the views, and all data is properly synced at all
times to the QJsonObject.

The available Themes are obtained from Repository::themes(), and
the json documents are obtained from each Theme::filePath(). All
build-in themes will be read-only, see Theme::isReadOnly().
Changing or deleting read-only themes is not possible. Instead,
changing a read-only theme first needs a copy of the read-only
theme.

Right now, this is a stand-alone cmake project, i.e. build in a
separate build folder and then start the `themedemo`.